Refactored map controls
