+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@types/classnames": "^2.2.6",
     "@types/leaflet": "^1.2.7",
+    "@types/leaflet-draw": "^0.4.13",
     "@types/proj4": "^2.3.4",
     "@types/prop-types": "^15.5.3",
     "@types/react": "^16.4.6",
@@ -30,12 +31,13 @@
     "html-webpack-plugin": "2.29.0",
     "jest": "22.4.2",
     "leaflet": "^1.3.1",
+    "leaflet-draw": "^1.0.2",
     "mobx": "^5.0.3",
     "mobx-react": "^5.2.3",
-    "proj4": "^2.4.4",
     "object-assign": "4.1.1",
     "postcss-flexbugs-fixes": "3.2.0",
     "postcss-loader": "2.0.8",
+    "proj4": "^2.4.4",
     "promise": "8.0.1",
     "raf": "3.4.0",
     "react": "^16.4.1",
