+++ src/components/map/coordinateControl.scss
@@ -1,0 +1,6 @@
+@import '../../common.scss';
+.container {
+  position: absolute;
+  right: 10px;
+  top: 10px;
+}
